requirements/base.txt
# Base
pytz==2018.7
python-slugify==1.2.6
Pillow==7.0.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==2.2.12

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.11.1
django-filter==2.4.0
django-cors-headers==3.5.0
drf-writable-nested==0.6.2


# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==4.1.2

# Celery
tornado==5.1.1
redis==3.5.3
django-redis==4.12.1
celery==4.4.7
flower==0.9.5
